Update this comparison based on Consul 1.5-1.6
I added a change, but the owners of this page will probably have a better edit than I did and can dismiss mine. I just wanted to shoot this pull request out so that this high-search-ranking page (#1 result when you search "Istio vs Consul") gets an update as soon as possible.